calculo de médias

+++ atividade04.py
@@ -1,8 +1,8 @@
 # Crie um programa que receba três notas de um aluno e calcule a média aritmética.
 
-nota1 = int(input(" digite a nota1"))
-nota2 = int(input(" digite a nota2"))
-nota3 = int(input(" digite a nota3"))
+nota1 = float(input(" digite a nota1"))
+nota2 = float(input(" digite a nota2"))
+nota3 = float(input(" digite a nota3"))
 
 media = nota1+nota2+nota3/3
